requirements.txt
asgiref==3.3.2
Django
pytz==2020.5
sqlparse==0.4.2
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
django-cors-headers
drf-yasg
drf-spectacular
inflection
jsonschema
jsonschema-specifications
packaging
pyYaml
referencing
rpds-py
typing-extensions
uritemplate
jwt
